feat: When Istio, serve model as a Knative Service
When using Istio the eirini app-ingress service is not deployed. Consequently
when creating a deployment for serving the model the Ingress is not created
to expose it.

In that way, deploy the model as a Knative Service which automatically takes
care of exposing the service through an Istio Ingress Gateway.