fix: Do not check for minio job when deploying mlflow
When deploying minio (as part of the mlflow helm chart) it starts a job
workload for creating the default bucket on minio. That job starts a pod
that is automatically deleted when succeeds. If the pod is deleted between
the checks (exists, running) the deployment fails as the pod does not exist
anymore when it tries to check if it is running.