refactor: Deploy tekton using tekton-operator
The tekton-operator is the quickest and easiest way to install,
upgrade and manage TektonCD Pipelines, Dashboard, Triggers on any
Kubernetes Cluster. Instead of maintaining one yaml for each tekton
component, with the tekton-operator we only need to maintain its
yaml and a yaml file for the deployment config (only a couple of lines).
Doing a upgrade is simple a metter of applying a the new yaml file
from the tekton-operator and it will handle updating all tekton
components.

This change also covers:
- Update takton components (pipelines(v0.23.0), triggers(v0.13.0),
  dashboard(v0.16.1)
- A new service account/roles with the minimal set of permissions
  required for tekton triggers (before it was a cluster-admin!)