feat: When installing an extension, register it in the registry
Extension registry is part of the core service, which is also
installed by the installer. So to access this registry, it's only
necessary to remember where fuseml-core service was installed.

Registering is done by POST calls to the registry, for building
the right payload it is necessary to build a structure that registry
expects. Until it is possible to just import fuseml-core module,
let's copy the necessary structs.