feat: Add script for downloading fuseml-installer
Add script for automatically downloading the correct `fuseml-installer`
binary according to the system it is being executed. By default the script
downloads the latest version and also upgrades the current binary if a new
version is available. If `TAG` is set it will download the binary from
the tag version.

This change also includes a CI job for testing the script/binary on
linux/macos/windows.

Fix: https://github.com/fuseml/fuseml/issues/207