feat: Added retries to HTTP calls
Network connection stability can vary, so it's good to have
a possibility to several retries whenever HTTP request is done.

Used retryablehttp library with the default values (4 retries).
Logging is turned off by default (so that debug messages are not
printed) and can be turned on by using "--verbose 1" argument to CLI.