fix: Wait for Pod to exist before checking the state
It could happen that when fuseml wants to check for state of a resource,
it does not exist yet, and the check fails.

So extend the wait by checking if the resource exist first.

For start, use this only for pods. It's also necesary for an extension
waitStep to provide a pod selector so that the wait can be executed.