fix: Map https port on registry (trow)
When pushing an image, kaniko performs a live check against the registry,
this check is first done through the https port and if there is no reply
it waits until the request times out then proceeds with http. This is
the case even when using the `--insecure` flag, resulting in
https://github.com/fuseml/fuseml/issues/174

To work around that, this commit maps the 443 port to the registry
container so when kaniko performs the live check the container can reply
with connection refused as there is no proccess listening on 443.

Fixes: #174