fix: retry when applying the tekton config
The CI is sometimes failing when installing FuseML/tekton as when
it created the tekton config the webhook for validating the config
is not ready yet.

This change adds a retry in case of failure creating the tekton config.
It will retry 3 times with an interval that should be enough for the
webhook to be ready.