ci: don't test the installer twice on release
The fuseml-installer test github action is currently triggered on
every possible push event on every branch: commits, but also tags.
This has the effect that this test is triggered twice when a release
tag is published: once on the actual tag push event and a second time
as a dependency in the release action tree.

Specifying the list of branches in the action trigger excludes tags
from the list of events that trigger this workflow. It will only be
executed on PR push and when commits are merged.